remove Type\UserLoader to main namespace and have it alias the base Message\User\Loader
